Fix the Github Action step
* Run RSpec instead of Rake default task

Less strict assertions, because ignorefiles are not always as minimal as theoretically possible.

* Fix failing spec if no ruby-vuln-db is present

The suite fails if there's no existing vulnerability db on the machine. stdout is not JSON parsable in this specific case, because it contains the download progress message as well. To fix this, we're downloading the database once before running the specs.

* Fix warning when running the specs

Remove dead code